Package renamed into more relevant one
