--- src/main/java/manufacturing/dao/CarDao.java
@@ -1,8 +1,0 @@
-package manufacturing.dao;
-
-import java.util.List;
-import manufacturing.model.Car;
-
-public interface CarDao extends AbstractDao<Car> {
-    List<Car> getAllByDriver(Long driverId);
-}
