--- src/main/java/manufacturing/dao/DriverDao.java
@@ -1,6 +1,0 @@
-package manufacturing.dao;
-
-import manufacturing.model.Driver;
-
-public interface DriverDao extends AbstractDao<Driver> {
-}
