--- src/main/java/manufacturing/dao/ManufacturerDao.java
@@ -1,6 +1,0 @@
-package manufacturing.dao;
-
-import manufacturing.model.Manufacturer;
-
-public interface ManufacturerDao extends AbstractDao<Manufacturer> {
-}
