--- src/main/java/manufacturing/dao/jdbc/ManufacturerDaoJdbc.java
@@ -1,116 +1,0 @@
-package manufacturing.dao.jdbc;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import manufacturing.dao.ManufacturerDao;
-import manufacturing.exception.DataProcessingException;
-import manufacturing.injections.Dao;
-import manufacturing.model.Manufacturer;
-import manufacturing.util.ConnectionUtil;
-
-@Dao
-public class ManufacturerDaoJdbc implements ManufacturerDao {
-
-    @Override
-    public Manufacturer create(Manufacturer manufacturer) {
-        String query = "INSERT INTO manufacturers (manufacturer_name, manufacturer_country) "
-                + "VALUES (?, ?)";
-        try (Connection connection = ConnectionUtil.getConnection();
-                PreparedStatement insertStatement =
-                        connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)) {
-
-            insertStatement.setString(1, manufacturer.getName());
-            insertStatement.setString(2, manufacturer.getCountry());
-            insertStatement.executeUpdate();
-            ResultSet resultSet = insertStatement.getGeneratedKeys();
-            if (resultSet.next()) {
-                manufacturer.setId(resultSet.getObject(1, Long.class));
-            }
-        } catch (SQLException e) {
-            throw new DataProcessingException("Could not insert " + manufacturer, e);
-        }
-        return manufacturer;
-    }
-
-    @Override
-    public Optional<Manufacturer> get(Long id) {
-        String query = "SELECT * FROM manufacturers WHERE manufacturer_id = ? AND deleted = FALSE";
-        Manufacturer manufacturer = null;
-        try (Connection connection = ConnectionUtil.getConnection();
-                PreparedStatement getByIdStatement = connection.prepareStatement(query)) {
-
-            getByIdStatement.setLong(1, id);
-            ResultSet resultSet = getByIdStatement.executeQuery();
-            if (resultSet.next()) {
-                manufacturer = createManufacturerInstance(resultSet);
-            }
-        } catch (SQLException e) {
-            throw new DataProcessingException("Could not get manufacturer by id " + id, e);
-        }
-        return Optional.ofNullable(manufacturer);
-    }
-
-    @Override
-    public List<Manufacturer> getAll() {
-        String query = "SELECT * FROM manufacturers WHERE deleted = FALSE";
-        List<Manufacturer> allManufacturers = new ArrayList<>();
-        try (Connection connection = ConnectionUtil.getConnection();
-                PreparedStatement getAllStatement = connection.prepareStatement(query)) {
-
-            ResultSet resultSet = getAllStatement.executeQuery();
-            while (resultSet.next()) {
-                allManufacturers.add(createManufacturerInstance(resultSet));
-            }
-        } catch (SQLException e) {
-            throw new DataProcessingException("Could not get all manufacturers", e);
-        }
-        return allManufacturers;
-    }
-
-    private Manufacturer createManufacturerInstance(ResultSet resultSet) throws SQLException {
-        Long manufacturerId = resultSet.getObject("manufacturer_id", Long.class);
-        String name = resultSet.getObject("manufacturer_name", String.class);
-        String country = resultSet.getObject("manufacturer_country", String.class);
-        Manufacturer manufacturer = new Manufacturer(name, country);
-        manufacturer.setId(manufacturerId);
-        return manufacturer;
-    }
-
-    @Override
-    public Manufacturer update(Manufacturer manufacturer) {
-        String query = "UPDATE manufacturers "
-                + "SET manufacturer_name = ?, manufacturer_country = ? "
-                + "WHERE manufacturer_id = ? AND deleted = FALSE";
-        try (Connection connection = ConnectionUtil.getConnection();
-                PreparedStatement updateStatement = connection.prepareStatement(query)) {
-
-            updateStatement.setString(1, manufacturer.getName());
-            updateStatement.setString(2, manufacturer.getCountry());
-            updateStatement.setLong(3, manufacturer.getId());
-            updateStatement.executeUpdate();
-        } catch (SQLException e) {
-            throw new DataProcessingException("Could not update " + manufacturer, e);
-        }
-        return manufacturer;
-    }
-
-    @Override
-    public boolean delete(Long id) {
-        String query = "UPDATE manufacturers SET deleted = true WHERE manufacturer_id = ?";
-        try (Connection connection = ConnectionUtil.getConnection();
-                PreparedStatement markDeletedStatement = connection.prepareStatement(query)) {
-
-            markDeletedStatement.setLong(1, id);
-            int deletedRows = markDeletedStatement.executeUpdate();
-            return deletedRows != 0;
-        } catch (SQLException e) {
-            throw new DataProcessingException("Could not delete manufacturer by id " + id, e);
-        }
-    }
-}
