--- src/main/java/manufacturing/exception/DataProcessingException.java
@@ -1,7 +1,0 @@
-package manufacturing.exception;
-
-public class DataProcessingException extends RuntimeException {
-    public DataProcessingException(String message, Throwable e) {
-        super(message, e);
-    }
-}
