--- src/main/java/manufacturing/injections/Dao.java
@@ -1,8 +1,0 @@
-package manufacturing.injections;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Dao {
-}
