--- src/main/java/manufacturing/model/Driver.java
@@ -1,63 +1,0 @@
-package manufacturing.model;
-
-import java.util.Objects;
-
-public class Driver {
-    private Long id;
-    private String name;
-    private String licenseNumber;
-
-    public Driver(String name, String licenseNumber) {
-        this.name = name;
-        this.licenseNumber = licenseNumber;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getLicenseNumber() {
-        return licenseNumber;
-    }
-
-    public void setLicenseNumber(String licenseNumber) {
-        this.licenseNumber = licenseNumber;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Driver driver = (Driver) o;
-        return name.equals(driver.name) && licenseNumber.equals(driver.licenseNumber);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name, licenseNumber);
-    }
-
-    @Override
-    public String toString() {
-        return "Driver{"
-                + "id=" + id
-                + ", name='" + name + '\''
-                + ", licenseNumber='" + licenseNumber + '\'' + '}';
-    }
-}
