--- src/main/java/manufacturing/model/Manufacturer.java
@@ -1,64 +1,0 @@
-package manufacturing.model;
-
-import java.util.Objects;
-
-public class Manufacturer {
-    private Long id;
-    private String name;
-    private String country;
-
-    public Manufacturer(String name, String country) {
-        this.name = name;
-        this.country = country;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Manufacturer that = (Manufacturer) o;
-        return name.equals(that.name) && country.equals(that.country);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name, country);
-    }
-
-    @Override
-    public String toString() {
-        return "Manufacturer{"
-                + "id=" + id
-                + ", name='" + name + '\''
-                + ", country='" + country + '\''
-                + '}';
-    }
-}
