--- src/main/java/manufacturing/service/DriverService.java
@@ -1,7 +1,0 @@
-package manufacturing.service;
-
-import manufacturing.model.Driver;
-
-public interface DriverService extends AbstractService<Driver> {
-
-}
