--- src/main/java/manufacturing/service/ManufacturerService.java
@@ -1,7 +1,0 @@
-package manufacturing.service;
-
-import manufacturing.model.Manufacturer;
-
-public interface ManufacturerService extends AbstractService<Manufacturer> {
-
-}
