--- src/main/java/manufacturing/service/impl/CarServiceImpl.java
@@ -1,59 +1,0 @@
-package manufacturing.service.impl;
-
-import java.util.List;
-import manufacturing.dao.CarDao;
-import manufacturing.injections.Inject;
-import manufacturing.injections.Service;
-import manufacturing.model.Car;
-import manufacturing.model.Driver;
-import manufacturing.service.CarService;
-
-@Service
-public class CarServiceImpl implements CarService {
-
-    @Inject
-    private CarDao carDao;
-
-    @Override
-    public Car create(Car car) {
-        return carDao.create(car);
-    }
-
-    @Override
-    public Car get(Long id) {
-        return carDao.get(id).orElseThrow(() ->
-                new RuntimeException("Car with id " + id + " not found"));
-    }
-
-    @Override
-    public List<Car> getAll() {
-        return carDao.getAll();
-    }
-
-    @Override
-    public Car update(Car car) {
-        return carDao.update(car);
-    }
-
-    @Override
-    public boolean delete(Long id) {
-        return carDao.delete(id);
-    }
-
-    @Override
-    public void addDriverToCar(Driver driver, Car car) {
-        car.getDrivers().add(driver);
-        carDao.update(car);
-    }
-
-    @Override
-    public void removeDriverFromCar(Driver driver, Car car) {
-        car.getDrivers().remove(driver);
-        carDao.update(car);
-    }
-
-    @Override
-    public List<Car> getAllByDriver(Long driverId) {
-        return carDao.getAllByDriver(driverId);
-    }
-}
