+++ src/main/java/taxi/dao/CarDao.java
@@ -1,0 +1,8 @@
+package taxi.dao;
+
+import java.util.List;
+import taxi.model.Car;
+
+public interface CarDao extends AbstractDao<Car> {
+    List<Car> getAllByDriver(Long driverId);
+}
