+++ src/main/java/taxi/dao/DriverDao.java
@@ -1,0 +1,6 @@
+package taxi.dao;
+
+import taxi.model.Driver;
+
+public interface DriverDao extends AbstractDao<Driver> {
+}
