+++ src/main/java/taxi/dao/ManufacturerDao.java
@@ -1,0 +1,6 @@
+package taxi.dao;
+
+import taxi.model.Manufacturer;
+
+public interface ManufacturerDao extends AbstractDao<Manufacturer> {
+}
