+++ src/main/java/taxi/exception/DataProcessingException.java
@@ -1,0 +1,7 @@
+package taxi.exception;
+
+public class DataProcessingException extends RuntimeException {
+    public DataProcessingException(String message, Throwable e) {
+        super(message, e);
+    }
+}
