+++ src/main/java/taxi/injections/Inject.java
@@ -1,0 +1,8 @@
+package taxi.injections;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Inject {
+}
