+++ src/main/java/taxi/service/DriverService.java
@@ -1,0 +1,7 @@
+package taxi.service;
+
+import taxi.model.Driver;
+
+public interface DriverService extends AbstractService<Driver> {
+
+}
