+++ src/main/java/taxi/service/ManufacturerService.java
@@ -1,0 +1,7 @@
+package taxi.service;
+
+import taxi.model.Manufacturer;
+
+public interface ManufacturerService extends AbstractService<Manufacturer> {
+
+}
